Refactor: Fix service layer has responsibility to get access token
서비스 레이어가 AccessToken 을 사용해서 API 를 실행하므로, AccessToken 과 관련된 작업을 Interactor 레이어 대신 서비스 레이어에서 하는게 낫다고 판단됨

+++ LabelLab/LabelLab/Service/GithubService.swift
@@ -9,30 +9,44 @@ import Foundation
 
 protocol GithubService {
     var session: URLSessionProtocol { get }
-    func requestRepositories(with token: String) async throws -> [GithubRepository]
-    func requestLabels(with token: String, of repository: GithubRepository) async throws -> [Label]
-    func createLabel(with token: String, to repository: GithubRepository, label: Label) async throws
-    func deleteLabel(with token: String, of repository: GithubRepository, label: Label) async throws
+    var accessTokenManager: AccessTokenManager { get }
+    func requestRepositories() async throws -> [GithubRepository]
+    func requestLabels(of repository: GithubRepository) async throws -> [Label]
+    func createLabel(to repository: GithubRepository, label: Label) async throws
+    func deleteLabel(of repository: GithubRepository, label: Label) async throws
+}
+
+extension GithubService {
+    var token: String? {
+        get async throws {
+            try await accessTokenManager.requestAccessToken()
+        }
+    }
 }
 
 final class RealGithubService {
     static let shared: GithubService = RealGithubService()
     let session: URLSessionProtocol
+    let accessTokenManager: AccessTokenManager
 
-    init(session: URLSessionProtocol = URLSession.shared) {
+    init(session: URLSessionProtocol = URLSession.shared, accessTokenManager: AccessTokenManager = RealAccessTokenManager()) {
         self.session = session
+        self.accessTokenManager = accessTokenManager
     }
 }
 
 // MARK: - Protocol 구현부
 extension RealGithubService: GithubService {
 
-    func requestLabels(with token: String, of repository: GithubRepository) async throws -> [Label] {
+    func requestLabels(of repository: GithubRepository) async throws -> [Label] {
+        guard let token = try await token else {
+            throw OAuthError.tokenNotExist
+        }
         let (data, _) = try await GithubRepAPI.labels(token: token, of: repository).request(with: session)
         return try JSONDecoder().decode([LabelDAO].self, from: data).map { $0.toLabel() }
     }
 
-    func createLabel(with token: String, to repository: GithubRepository, label: Label) async throws {
+    func createLabel(to repository: GithubRepository, label: Label) async throws {
         struct LabelUpload: Encodable {
             let name: String
             let description: String
@@ -44,6 +58,10 @@ extension RealGithubService: GithubService {
                 self.color = label.hex
             }
         }
+
+        guard let token = try await token else {
+            throw OAuthError.tokenNotExist
+        }
         let labelData = try JSONEncoder().encode(LabelUpload(from: label))
         let (_, response) = try await GithubRepAPI.createLabel(token: token, of: repository, label: label).post(with: session, data: labelData, delegate: nil)
         guard let response = response as? HTTPURLResponse, response.statusCode == 201 else {
@@ -51,14 +69,20 @@ extension RealGithubService: GithubService {
         }
     }
 
-    func deleteLabel(with token: String, of repository: GithubRepository, label: Label) async throws {
+    func deleteLabel(of repository: GithubRepository, label: Label) async throws {
+        guard let token = try await token else {
+            throw OAuthError.tokenNotExist
+        }
         let (_, response) = try await GithubRepAPI.deleteLabel(token: token, of: repository, label: label).request(with: session)
         guard let response = response as? HTTPURLResponse, response.statusCode == 204 else {
             throw URLError(.badServerResponse)
         }
     }
 
-    func requestRepositories(with token: String) async throws -> [GithubRepository] {
+    func requestRepositories() async throws -> [GithubRepository] {
+        guard let token = try await token else {
+            throw OAuthError.tokenNotExist
+        }
         let (data, _) = try await GithubRepAPI.repositories(token: token).request(with: session)
         let repositories: [GithubRepository] = try JSONDecoder().decode([GithubRepository].self, from: data)
